refactor(ratchet): route ratchet/exclude emitters through writeJsonManifest
The no-explicit-any, test-fixture-parity, and standalone-associations
generators still emitted raw `JSON.stringify(x, null, 2)`. Their tracked
outputs pass `prettier --check` only by coincidence — flat arrays of long
path strings prettier won't collapse — so the first short-enough entry would
churn the tree on every regeneration, the trap #4987 fixed for manifests.

Route all three through `writeJsonManifest()` so they share the single
prettier formatting code path. Convert the no-explicit-any generator from
`.mjs` (plain node, node:* imports) to `.ts` run via tsx so it can import the
shared helper. Committed outputs are already prettier-stable, so no reformat
results; verified writeJsonManifest reproduces the committed bytes
byte-identically. Extend the `manifest emitters` source guard to cover them.
