+++ scripts/api-compare/write-json-manifest.test.ts
@@ -105,4 +105,24 @@ describe("manifest emitters", () => {
       );
     }
   });
+
+  // The ratchet/exclude generators share the same churn trap as the manifests:
+  // their outputs pass `prettier --check` today only because the current entries
+  // are long path strings prettier won't collapse. Routing them through
+  // writeJsonManifest keeps the single formatting code path; guard against a
+  // regression back to raw `JSON.stringify` + writeFileSync.
+  it("ratchet/exclude generators emit via writeJsonManifest", () => {
+    const generators = [
+      "generate-no-explicit-any-allowlist.ts",
+      "generate-fixture-parity-exclude.ts",
+      "generate-standalone-associations-exclude.ts",
+    ];
+    for (const name of generators) {
+      const src = fs.readFileSync(path.join(REPO_ROOT, "scripts", name), "utf8");
+      expect(src, `${name} must import writeJsonManifest`).toMatch(/\bwriteJsonManifest\b/);
+      expect(src, `${name} must not write JSON with writeFileSync`).not.toMatch(
+        /writeFileSync\s*\([^)]*JSON\.stringify/,
+      );
+    }
+  });
 });
